+++ modules/main/src/main/java/com/nisovin/shopkeepers/SKShopkeepersPlugin.java
@@ -30,6 +30,7 @@ import com.nisovin.shopkeepers.config.lib.ConfigLoadException;
 import com.nisovin.shopkeepers.container.protection.ProtectedContainers;
 import com.nisovin.shopkeepers.container.protection.RemoveShopOnContainerBreak;
 import com.nisovin.shopkeepers.debug.Debug;
+import com.nisovin.shopkeepers.debug.DebugOptions;
 import com.nisovin.shopkeepers.debug.events.EventDebugger;
 import com.nisovin.shopkeepers.debug.trades.TradingCountListener;
 import com.nisovin.shopkeepers.dependencies.worldguard.WorldGuardDependency;
@@ -592,6 +593,29 @@ public class SKShopkeepersPlugin extends JavaPlugin implements InternalShopkeepe
 	@Override
 	public SKShopkeeperStorage getShopkeeperStorage() {
 		return shopkeeperStorage;
+	}
+
+	// ITEM UPDATES
+
+	@Override
+	public int updateItems() {
+		Log.debug(DebugOptions.itemUpdates, "Updating all items.");
+
+		// Note: Not safe to be called from inside inventory events!
+		uiRegistry.abortUISessions();
+
+		int updatedItems = Settings.getInstance().updateItems();
+
+		int shopkeeperUpdatedItems = 0;
+		for (AbstractShopkeeper shopkeeper : shopkeeperRegistry.getAllShopkeepers()) {
+			shopkeeperUpdatedItems += shopkeeper.updateItems();
+		}
+		if (shopkeeperUpdatedItems > 0) {
+			updatedItems += shopkeeperUpdatedItems;
+			shopkeeperStorage.save();
+		}
+
+		return updatedItems;
 	}
 
 	// COMMANDS