+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/player/sell/SKSellingPlayerShopkeeper.java
@@ -23,6 +23,7 @@ import com.nisovin.shopkeepers.shopkeeper.migration.ShopkeeperDataMigrator;
 import com.nisovin.shopkeepers.shopkeeper.offers.SKPriceOffer;
 import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopkeeper;
 import com.nisovin.shopkeepers.util.annotations.ReadOnly;
+import com.nisovin.shopkeepers.util.annotations.ReadWrite;
 import com.nisovin.shopkeepers.util.data.property.BasicProperty;
 import com.nisovin.shopkeepers.util.data.property.Property;
 import com.nisovin.shopkeepers.util.data.serialization.InvalidDataException;
@@ -30,6 +31,7 @@ import com.nisovin.shopkeepers.util.inventory.InventoryUtils;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 import com.nisovin.shopkeepers.util.java.CollectionUtils;
 import com.nisovin.shopkeepers.util.java.Validate;
+import com.nisovin.shopkeepers.util.logging.Log;
 
 public class SKSellingPlayerShopkeeper
 		extends AbstractPlayerShopkeeper implements SellingPlayerShopkeeper {
@@ -68,6 +70,31 @@ public class SKSellingPlayerShopkeeper
 		super.saveDynamicState(shopkeeperData, saveAll);
 		this.saveOffers(shopkeeperData);
 	}
+
+	// ITEM UPDATES
+
+	@Override
+	protected int updateItems(String logPrefix, @ReadWrite ShopkeeperData shopkeeperData) {
+		int updatedItems = super.updateItems(logPrefix, shopkeeperData);
+		updatedItems += updateOfferItems(logPrefix, shopkeeperData);
+		return updatedItems;
+	}
+
+	private static int updateOfferItems(String logPrefix, @ReadWrite ShopkeeperData shopkeeperData) {
+		try {
+			var updatedOffers = new ArrayList<PriceOffer>(shopkeeperData.get(OFFERS));
+			var updatedItems = SKPriceOffer.updateItems(updatedOffers, logPrefix);
+			if (updatedItems > 0) {
+				shopkeeperData.set(OFFERS, updatedOffers);
+				return updatedItems;
+			}
+		} catch (InvalidDataException e) {
+			Log.warning(logPrefix + "Failed to load '" + OFFERS.getName() + "'!", e);
+		}
+		return 0;
+	}
+
+	//
 
 	@Override
 	public SellingPlayerShopType getType() {